+++ lib/src/type_definitions.dart
@@ -4,3 +4,4 @@ import 'package:firebase_auth/firebase_auth.dart';
 
 typedef OnLoginSuccess = FutureOr<void> Function(UserCredential, bool);
 typedef OnLoginFailed = FutureOr<void> Function(FirebaseAuthException);
+typedef OnError = FutureOr<void> Function(Object);